updated user

+++ app/src/main/java/com/mobdeve/tutorfinderapp/User.java
@@ -28,7 +28,7 @@ public class User {
         this.categories = categories;
     }
 
-    public double getFee() {
+    public float getFee() {
         return fee;
     }
